% Auto-generated: do not edit by hand
\name{dashRenderjson}

\alias{dashRenderjson}

\title{DashRenderjson component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dashRenderjson(id=NULL, data=NULL, max_depth=NULL, theme=NULL,
invert_theme=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{data}{Named list. The value displayed in the input.}

\item{max_depth}{Numeric. Dash-assigned callback that should be called to report property changes
to Dash, to make them available for callbacks.}

\item{theme}{Named list. }

\item{invert_theme}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

